requirement.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
